pyproject.toml
[project]
name = "day-8"
version = "0.1.0"
description = ""
authors = [
    {name = "oefremidis",email = "78813098+oefremidis@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12, <3.15"
 
[tool.poetry]
packages = [{include = "day_8", from = "src"}]

[tool.poetry.dependencies]
langchain-openai = "^0.3.35"
langchain-core = "^0.3.79"
langgraph = "^0.6.10"
dotenv = "^0.9.9"
langchain-community = "^0.3.31"
ddgs = "^9.6.1"
langchain-huggingface = "^0.3.1"
huggingface-hub = "^0.35.3"
pydantic = "^2.12.2"
chromadb = "^1.1.1"
neo4j = "^6.0.2"
transformers = "^4.57.1"
torch = "^2.9.0"
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


